@@ -13,24 +13,24 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.7",
+    "@angular/animations": "^18.0.2",
     "@angular/cdk": "^17.3.7",
-    "@angular/common": "^17.3.7",
-    "@angular/compiler": "^17.3.7",
-    "@angular/core": "^17.3.7",
-    "@angular/forms": "^17.3.7",
-    "@angular/platform-browser": "^17.3.7",
-    "@angular/platform-browser-dynamic": "^17.3.7",
-    "@angular/router": "^17.3.7",
+    "@angular/common": "^18.0.2",
+    "@angular/compiler": "^18.0.2",
+    "@angular/core": "^18.0.2",
+    "@angular/forms": "^18.0.2",
+    "@angular/platform-browser": "^18.0.2",
+    "@angular/platform-browser-dynamic": "^18.0.2",
+    "@angular/router": "^18.0.2",
     "moment-mini": "^2.24.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.14.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.6",
-    "@angular/cli": "~17.3.6",
-    "@angular/compiler-cli": "^17.3.7",
+    "@angular-devkit/build-angular": "^18.0.3",
+    "@angular/cli": "~18.0.3",
+    "@angular/compiler-cli": "^18.0.2",
     "@types/jasmine": "~4.3.0",
     "jasmine-core": "~4.6.0",
     "karma": "~6.4.0",
@@ -38,7 +38,7 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "ng-packagr": "^17.3.0",
+    "ng-packagr": "^18.0.0",
     "prettier": "^3.2.5",
     "prettier-plugin-jsdoc": "^1.3.0",
     "prettier-plugin-organize-imports": "^3.2.4",
