@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { ItskGridModule } from '../itsk-grid.module';
 import { BasicComponent } from './basic/basic.component';
@@ -7,9 +7,6 @@ import { CustomComponentsComponent } from './custom-components/custom-components
 import { GroupingGridComponent } from './grouping-grid/grouping-grid.component';
 import { TreeGridComponent } from './tree-grid/tree-grid.component';
 
-@NgModule({
-  declarations: [BasicComponent, GroupingGridComponent, TreeGridComponent, CustomComponentsComponent],
-  exports: [BasicComponent, GroupingGridComponent, TreeGridComponent, CustomComponentsComponent],
-  imports: [CommonModule, ItskGridModule, HttpClientModule],
-})
+@NgModule({ declarations: [BasicComponent, GroupingGridComponent, TreeGridComponent, CustomComponentsComponent],
+    exports: [BasicComponent, GroupingGridComponent, TreeGridComponent, CustomComponentsComponent], imports: [CommonModule, ItskGridModule], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class GridSamplesModule {}
